fix: run the legacy generations backfill exactly once, ever
backfill_from_generations ran on every ApplicationStore::open and only
skipped a row once application_id was set or created_at crossed the
epoch guard. Deleting an Application detaches its generation rows back
to application_id=NULL without touching created_at, so a genuinely
pre-epoch row stayed pre-epoch and unlinked forever -- the next boot's
unconditional re-scan recreated the Application the user just deleted.
The epoch guard narrowed that window; it never closed it.

Add a durable backfill_state marker (new append-only migration) and
check it first: once the one-time legacy backfill has completed for a
data dir, later boots skip it entirely. link_orphaned_generations stays
safe to re-run every boot since it is lookup-only and can never create
an Application.

Split the backfill + its marker out of applications/mod.rs into
migrations.rs (same store, same pattern as reminders.rs/contact.rs) to
stay under the R8 LOC cap, and allowlist that file for R3's rusqlite
confinement rule, mirroring the existing per-domain-store entries.
