+++ apps/desktop/src-tauri/src/applications/migrations.rs
@@ -9,7 +9,15 @@
 //! Every migration is additive and forward-safe — `ADD COLUMN` with a default, or
 //! a guarded backfill `UPDATE`. Nothing here drops a column or a table.
 
-use crate::db::Migration;
+use std::path::Path;
+
+use rusqlite::params;
+
+use crate::ai_generations::ApplicationAnswer;
+use crate::db::{now_ms, ts_from_db, ts_to_db, Migration};
+use crate::error::AppResult;
+
+use super::{normalize_job_url, ApplicationMeta, ApplicationStore, APPLICATIONS_FEATURE_EPOCH_MS};
 
 /// The whitespace set SQLite's two-argument `TRIM(x, y)` must strip so the
 /// `unify_application_contact` emptiness test agrees with Rust's `str::trim` —
@@ -204,4 +212,170 @@ pub(super) const MIGRATIONS: &[Migration] = &[
             ))
         },
     },
+    Migration {
+        name: "add_backfill_state",
+        up: |conn| {
+            // Durable one-shot marker table for `ApplicationStore::
+            // backfill_from_generations` below — a name/done_at row means that
+            // migration has already run and must never re-scan
+            // `ai_generations.db` again. See the constant's own doc for why a
+            // per-row epoch guard alone cannot close the resurrection window.
+            conn.execute_batch(
+                "CREATE TABLE IF NOT EXISTS backfill_state (
+                    name    TEXT PRIMARY KEY,
+                    done_at INTEGER NOT NULL
+                );",
+            )
+        },
+    },
 ];
+
+/// `backfill_state.name` for [`ApplicationStore::backfill_from_generations`]'s
+/// durable one-shot marker: once set, [`super::ApplicationStore::open`] never
+/// re-scans `ai_generations.db` again — see that function's own doc for why
+/// re-running it resurrects a deletion.
+const LEGACY_BACKFILL_MARKER: &str = "legacy_generations_backfill";
+
+impl ApplicationStore {
+    /// **One-time**, forward-safe backfill from the pre-split `ai_generations.db`:
+    ///
+    /// 1. `ALTER TABLE ai_generations ADD COLUMN application_id` (+ index) — guarded
+    ///    by [`crate::db::column_exists`] so a re-run is a no-op.
+    /// 2. For each generation OLDER than [`APPLICATIONS_FEATURE_EPOCH_MS`] (own
+    ///    doc; younger left NULL), ensure exactly one `Application(status=
+    ///    applied, applied_at=created_at)` keyed by `normalize(job_url)`,
+    ///    link the gen's `application_id`, and seed one status event.
+    ///
+    /// **Runs at most once per data dir, ever — a migration, not a recurring
+    /// reconciliation.** Every row it repairs predates the Applications
+    /// feature entirely, so once it has run there is nothing left it could
+    /// legitimately still find. Re-running it on a later boot does not merely
+    /// waste work: a row still unlinked at that point is usually one whose
+    /// Application the user deliberately deleted since — deleting an
+    /// Application detaches its generations back to `NULL` without touching
+    /// `created_at`, so the row is STILL pre-epoch and STILL unmatched, and an
+    /// unconditional re-scan would recreate the Application the delete just
+    /// removed. [`APPLICATIONS_FEATURE_EPOCH_MS`] alone narrows that window;
+    /// it cannot close it (own doc there).
+    ///
+    /// Guarded by [`LEGACY_BACKFILL_MARKER`]: checked FIRST (an early return,
+    /// no I/O against `ai_generations.db` at all once set) and set LAST, only
+    /// after every row above has been processed — so a run that errors out
+    /// partway is retried on the next boot instead of silently marked done.
+    pub(super) fn backfill_from_generations(&self, data_dir: &Path) -> AppResult<()> {
+        if self.legacy_backfill_done() {
+            return Ok(()); // ran once already — see doc above, never re-scan
+        }
+        let gen_path = data_dir.join("ai_generations.db");
+        if gen_path.exists() {
+            let gen_conn = crate::db::open(&gen_path)?;
+
+            // Step 1: add the FK column + index if absent (idempotent guard).
+            if !crate::db::column_exists(&gen_conn, "ai_generations", "application_id") {
+                gen_conn.execute_batch(
+                    "ALTER TABLE ai_generations ADD COLUMN application_id TEXT;
+                     CREATE INDEX IF NOT EXISTS idx_ai_generations_application_id
+                         ON ai_generations(application_id);",
+                )?;
+            }
+
+            struct GenRow {
+                id: String,
+                created_at: u64,
+                candidate: String,
+                title: String,
+                company: String,
+                job_url: String,
+                board: String,
+                answers: Vec<ApplicationAnswer>,
+                brief: String,
+                application_id: Option<String>,
+            }
+
+            let mut stmt = gen_conn.prepare(
+                "SELECT id, created_at, candidate_name, job_title, company_name,
+                        job_url, board, application_answers, company_brief, application_id
+                 FROM ai_generations
+                 ORDER BY created_at ASC",
+            )?;
+            let rows: Vec<GenRow> = stmt
+                .query_map([], |row| {
+                    let answers_json: String = row.get(7)?;
+                    let app_id: Option<String> = row.get(9).unwrap_or(None);
+                    Ok(GenRow {
+                        id: row.get(0)?,
+                        created_at: ts_from_db(row.get::<_, i64>(1)?),
+                        candidate: row.get(2)?,
+                        title: row.get(3)?,
+                        company: row.get(4)?,
+                        job_url: row.get(5)?,
+                        board: row.get(6)?,
+                        answers: serde_json::from_str(&answers_json).unwrap_or_default(),
+                        brief: row.get(8)?,
+                        application_id: app_id.filter(|s| !s.is_empty()),
+                    })
+                })?
+                .filter_map(|r| r.ok())
+                .collect();
+            drop(stmt);
+
+            for g in rows {
+                if g.application_id.is_some() {
+                    continue; // already linked — idempotent skip
+                }
+                if g.created_at >= APPLICATIONS_FEATURE_EPOCH_MS {
+                    continue; // modern + still unmatched — own doc on the constant
+                }
+                let normalized = normalize_job_url(&g.job_url);
+                let meta = ApplicationMeta {
+                    company: g.company,
+                    title: g.title,
+                    candidate: g.candidate,
+                    brief: g.brief,
+                    job_description: String::new(), // ponytail: legacy generations carry no JD column
+                    answers: g.answers,
+                    job_summary: String::new(),
+                    salary_min: None,
+                    salary_max: None,
+                    salary_currency: None,
+                };
+                // Empty url → a fresh per-gen Application (no shared key to merge on);
+                // non-empty url → merge into that url's Application if one exists.
+                let app_id = self.upsert_internal(
+                    &normalized,
+                    &g.board,
+                    &meta,
+                    &meta.job_description,
+                    Some(g.created_at),
+                )?;
+                gen_conn.execute(
+                    "UPDATE ai_generations SET application_id = ?2 WHERE id = ?1",
+                    params![g.id, app_id],
+                )?;
+            }
+        } // else: fresh install — still marked done below (nothing to find, ever).
+
+        self.mark_legacy_backfill_done()
+    }
+
+    /// Whether the one-shot marker [`LEGACY_BACKFILL_MARKER`] is set.
+    fn legacy_backfill_done(&self) -> bool {
+        let conn = self.conn.lock();
+        conn.query_row(
+            "SELECT 1 FROM backfill_state WHERE name = ?1",
+            params![LEGACY_BACKFILL_MARKER],
+            |_| Ok(()),
+        )
+        .is_ok()
+    }
+
+    /// Set it. `INSERT OR REPLACE` so a caller never special-cases "already set".
+    fn mark_legacy_backfill_done(&self) -> AppResult<()> {
+        let conn = self.conn.lock();
+        conn.execute(
+            "INSERT OR REPLACE INTO backfill_state (name, done_at) VALUES (?1, ?2)",
+            params![LEGACY_BACKFILL_MARKER, ts_to_db(now_ms())],
+        )?;
+        Ok(())
+    }
+}
